feat: middleware admin, fix: layouts app

+++ app/Http/Controllers/DivisionCOntroller.php
@@ -1,4 +1,5 @@
 <?php
+
 namespace App\Http\Controllers;
 
 use App\Models\Division;
@@ -8,17 +9,26 @@ class DivisionController extends Controller
 {
     public function index()
     {
+        if (auth()->check() && auth()->user()->role !== 'admin') {
+            abort(403, 'Unauthorized action.');
+        }
         $divisions = Division::all();
         return view('admin.pages.divisions.index', compact('divisions'));
     }
 
     public function create()
     {
+        if (auth()->check() && auth()->user()->role !== 'admin') {
+            abort(403, 'Unauthorized action.');
+        }
         return view('admin.pages.divisions.create');
     }
 
     public function store(Request $request)
     {
+        if (auth()->check() && auth()->user()->role !== 'admin') {
+            abort(403, 'Unauthorized action.');
+        }
         $request->validate([
             'name' => 'required|string|max:255',
         ]);
@@ -30,11 +40,17 @@ class DivisionController extends Controller
 
     public function edit(Division $division)
     {
+        if (auth()->check() && auth()->user()->role !== 'admin') {
+            abort(403, 'Unauthorized action.');
+        }
         return view('admin.pages.divisions.edit', compact('division'));
     }
 
     public function update(Request $request, Division $division)
     {
+        if (auth()->check() && auth()->user()->role !== 'admin') {
+            abort(403, 'Unauthorized action.');
+        }
         $request->validate([
             'name' => 'required|string|max:255',
         ]);
@@ -46,6 +62,9 @@ class DivisionController extends Controller
 
     public function destroy(Division $division)
     {
+        if (auth()->check() && auth()->user()->role !== 'admin') {
+            abort(403, 'Unauthorized action.');
+        }
         $division->delete();
 
         return redirect()->route('divisions.index')->with('success', 'Division deleted successfully.');
